requirements.txt
mysqlx == 2.1.3
mysqlx-connector-python == 8.3.0
Levenshtein == 0.25.1